feat: update migration script
