+++ sample/annotationless_perception/scenario.ja.yaml
@@ -20,4 +20,3 @@ Evaluation:
   Datasets:
     - sample_dataset:
         VehicleId: default
-
